# Avoid boxing in Framing