--- a/actor/src/main/scala-2.12/org/apache/pekko/util/ByteString.scala
+++ b/actor/src/main/scala-2.12/org/apache/pekko/util/ByteString.scala
@@ -710,8 +710,7 @@ object ByteString {
             } else {
               val subIndexOf = bs.indexOf(elem, relativeIndex)
               if (subIndexOf < 0) {
-                val nextString = bsIdx + 1
-                find(nextString, relativeIndex - bs.length, bytesPassed + bs.length)
+                find(bsIdx + 1, relativeIndex - bs.length, bytesPassed + bs.length)
               } else subIndexOf + bytesPassed
             }
           }
@@ -738,8 +737,7 @@ object ByteString {
             } else {
               val subIndexOf = bs.indexOf(elem, relativeIndex)
               if (subIndexOf < 0) {
-                val nextString = bsIdx + 1
-                find(nextString, relativeIndex - bs.length, bytesPassed + bs.length)
+                find(bsIdx + 1, relativeIndex - bs.length, bytesPassed + bs.length)
               } else subIndexOf + bytesPassed
             }
           }