--- a/actor/src/main/scala-2.12/org/apache/pekko/util/ByteString.scala
+++ b/actor/src/main/scala-2.12/org/apache/pekko/util/ByteString.scala
@@ -841,6 +841,9 @@ sealed abstract class ByteString extends IndexedSeq[Byte] with IndexedSeqOptimiz
 
   override def indexWhere(p: Byte => Boolean): Int = iterator.indexWhere(p)
 
+  // optimized in subclasses
+  override def indexOf[B >: Byte](elem: B): Int = indexOf[B](elem, 0)
+
   // optimized version of indexOf for bytes, implemented in subclasses
   /**
    * Finds index of first occurrence of some byte in this ByteString after or at some start index.
@@ -853,7 +856,7 @@ sealed abstract class ByteString extends IndexedSeq[Byte] with IndexedSeqOptimiz
    *           to `elem`, or `-1`, if none exists.
    *  @since 1.1.0
    */
-  def indexOf(elem: Byte, from: Int): Int
+  def indexOf(elem: Byte, from: Int): Int = indexOf[Byte](elem, from)
 
   /**
    * Finds index of first occurrence of some byte in this ByteString.