--- a/actor/src/main/scala-2.13/org/apache/pekko/util/ByteString.scala
+++ b/actor/src/main/scala-2.13/org/apache/pekko/util/ByteString.scala
@@ -717,8 +717,7 @@ object ByteString {
             } else {
               val subIndexOf = bs.indexOf(elem, relativeIndex)
               if (subIndexOf < 0) {
-                val nextString = bsIdx + 1
-                find(nextString, relativeIndex - bs.length, bytesPassed + bs.length)
+                find(bsIdx + 1, relativeIndex - bs.length, bytesPassed + bs.length)
               } else subIndexOf + bytesPassed
             }
           }
@@ -744,8 +743,7 @@ object ByteString {
             } else {
               val subIndexOf = bs.indexOf(elem, relativeIndex)
               if (subIndexOf < 0) {
-                val nextString = bsIdx + 1
-                find(nextString, relativeIndex - bs.length, bytesPassed + bs.length)
+                find(bsIdx + 1, relativeIndex - bs.length, bytesPassed + bs.length)
               } else subIndexOf + bytesPassed
             }
           }