--- a/actor/src/main/scala-3/org/apache/pekko/util/ByteString.scala
+++ b/actor/src/main/scala-3/org/apache/pekko/util/ByteString.scala
@@ -244,6 +244,19 @@ object ByteString {
       }
     }
 
+    override def indexOfByte(elem: Byte, from: Int): Int = {
+      if (from >= length) -1
+      else {
+        var found = -1
+        var i = math.max(from, 0)
+        while (i < length && found == -1) {
+          if (bytes(i) == elem) found = i
+          i += 1
+        }
+        found
+      }
+    }
+
     override def slice(from: Int, until: Int): ByteString =
       if (from <= 0 && until >= length) this
       else if (from >= length || until <= 0 || from >= until) ByteString.empty
@@ -423,6 +436,19 @@ object ByteString {
     }
 
     override def indexOf[B >: Byte](elem: B, from: Int): Int = {
+      if (from >= length) -1
+      else {
+        var found = -1
+        var i = math.max(from, 0)
+        while (i < length && found == -1) {
+          if (bytes(startIndex + i) == elem) found = i
+          i += 1
+        }
+        found
+      }
+    }
+
+    override def indexOfByte(elem: Byte, from: Int): Int = {
       if (from >= length) -1
       else {
         var found = -1
@@ -703,6 +729,33 @@ object ByteString {
       }
     }
 
+    override def indexOfByte(elem: Byte, from: Int): Int = {
+      if (from >= length) -1
+      else {
+        val byteStringsSize = bytestrings.size
+
+        @tailrec
+        def find(bsIdx: Int, relativeIndex: Int, bytesPassed: Int): Int = {
+          if (bsIdx >= byteStringsSize) -1
+          else {
+            val bs = bytestrings(bsIdx)
+
+            if (bs.length <= relativeIndex) {
+              find(bsIdx + 1, relativeIndex - bs.length, bytesPassed + bs.length)
+            } else {
+              val subIndexOf = bs.indexOfByte(elem, relativeIndex)
+              if (subIndexOf < 0) {
+                val nextString = bsIdx + 1
+                find(nextString, relativeIndex - bs.length, bytesPassed + bs.length)
+              } else subIndexOf + bytesPassed
+            }
+          }
+        }
+
+        find(0, math.max(from, 0), 0)
+      }
+    }
+
     override def copyToArray[B >: Byte](dest: Array[B], start: Int, len: Int): Int = {
       if (bytestrings.size == 1) bytestrings.head.copyToArray(dest, start, len)
       else {
@@ -824,6 +877,28 @@ sealed abstract class ByteString
 
   // optimized in subclasses
   override def indexOf[B >: Byte](elem: B, from: Int): Int = indexOf(elem, from)
+
+  // optimized version of indexOf for bytes, optimized in subclasses
+  /** Finds index of first occurrence of some byte in this ByteString after or at some start index.
+    *
+    * Similar to indexOf, but it avoids boxing if the value is already a byte.
+    *
+    *  @param   elem   the element value to search for.
+    *  @param   from   the start index
+    *  @return  the index `>= from` of the first element of this ByteString that is equal (as determined by `==`)
+    *           to `elem`, or `-1`, if none exists.
+    */
+  def indexOfByte(elem: Byte, from: Int): Int = indexOf(elem, from)
+
+  /** Finds index of first occurrence of some byte in this ByteString.
+    *
+    * Similar to indexOf, but it avoids boxing if the value is already a byte.
+    *
+    *  @param   elem   the element value to search for.
+    *  @return  the index `>= from` of the first element of this ByteString that is equal (as determined by `==`)
+    *           to `elem`, or `-1`, if none exists.
+    */
+  def indexOfByte(elem: Byte): Int = indexOfByte(elem, 0)
 
   override def grouped(size: Int): Iterator[ByteString] = {
     if (size <= 0) {