… actor-tests/src/test/scala/org/apache/pekko/util/ByteStringSpec.scala
… bench-jmh/src/main/scala/org/apache/pekko/stream/FramingBenchmark.scala
… stream/src/main/scala/org/apache/pekko/stream/scaladsl/Framing.scala