@He-Pin 2024-04-04
The byte searching can do with SIMD too, but I lack of time to do this:(