@He-Pin 2024-04-15
I think we can delegate the `indexOf[B >: Byte](elem: B, from: Int): Int` to `indexOf(elem: Byte, from: Int): Int` with `Byte.unbox`, wdyt, the reduce the duplicated code. @JD557