@JD557 2024-04-15
> That still feels like a footgun, since it is quite unclear in which situations the new overload will be selected? Should we document in which case the static overload resolution will choose the new method?

I don't have any strong feelings about the naming. I used this scheme because it's the same that's used by ByteIterator (See: https://github.com/apache/pekko/blob/3cd08016b57e95f9244f0990282dd956c173b82e/actor/src/main/scala-3/org/apache/pekko/util/ByteIterator.scala#L498-L502).

If we change the method here (e.g. to `indexOfByte`, like in 18c5db1), then I think it should also be changed in `ByteIterator`.

However, on that note:

> Is it somehow possible to avoid the massive code duplication here (probably hard since the whole existing scheme is built around the function of Any.==)? If not it should be documented.

Looks like `ByteIterator` avoids the duplication by delegating the comparison to `indexWhere`. Not sure if the lambda could have a performance impact, though. I would need to benchmark this again.