@He-Pin 2024-04-15
@JD557 I think we can delegate the current version to the specialized version, even it can happen in type theory but will not happen in real use case if user don't try to do strange thing.