@JD557 2024-04-15
So, I was doing some more tests with a specialized version of `indexWhere`, but that's noticeably slower (even though it's faster than `ByteIterator#indexWhere`). So I don't see many ways to reduce the duplication.

As such, I'm not sure how to proceed with this PR:
- Should I introduce a specialized `indexWhere` anyway, leading to code multiplication?
- Do I keep the `indexOf` name (like the `ByteIterator` does) or do I name it something else?