@He-Pin 2024-04-15
@JD557 I think if you want to change the method name, you can using the name of firstIndexOf.

And for the api, because it's using the scala collection's indexOf，so the input B where is actually a Byte, but now can be a any.

As for selection, I think the compiler will choose this new method when it know the elem for testing is a type Byte but will not when It doesn't. 

As for the old index Of, which will always return -1 us the elem is not a Number/can boxed to Byte, so a quick check and delegate to the specified one will not harm too much. Who will using a ByteString to indexOf a any?