@JD557 2024-04-15
I don't think that `BoxesRunTime` trick is enough, unfortunately.

Say one asks `ByteString.indexOf(500)`. This should obviously return `-1`, as it is impossible for a byte to have that value.
However `BoxesRunTime.toByte(500)` returns `-12: Byte`.

Ideally we would check if the number is between `Byte.MinValue` and `Byte.MaxValue` before boxing, but even that is not enough, due to floating point values (BoxesRunTime.toByte(127.9) == 127.toByte == 127.0).

Although I imagine most problems would come from `Int` and `Char`.