@He-Pin 2024-04-15
Another thing is , when we using code like `indexOf(`a`, 0)` the new method will not be selected too :(