@JD557 2024-04-20
I can try to fix it, but out of curiosity, isn't it OK to have MiMa issues, since this only targets 1.1.x?

I was ignoring the issue because I thought 1.1.0 was going to break bincompat with 1.0.x anyway