@sirthias 2024-04-25
The only comment I have is that there would be a potentially _much_ more efficient implementation of `indexOf(elem: Byte, from: Int)` if we somehow had low-level (unsafe) access to the byte array and could read 8 bytes at once into a `long`.
Then we could use SWAR (SIMD within a register) and reduce the loop count by 8 (asymptotically) over the current implementation.
Doesn't pekko already have an `Unsafe` access construct somewhere?
IIRC Akka did...