@sirthias 2024-04-25
Yes, there is a (much more complicated) SWAR loop implemented in [borer's JSON parser](https://github.com/sirthias/borer/blob/4d639de35ade1bc2513a50a9da9587451b322476/core/src/main/scala/io/bullet/borer/json/JsonParser.scala#L430).
Here we only have to look for a single known byte rather than a whole set of different characters and we also don't have to copy segments and do UTF8 decoding at the same time.

But the whole thing only makes sense if we can really get down to raw byte access via `Unsafe` or more modern means. And on ScalaJS a SWAR approach will just create overhead and be a lot slower than the simple loop.