@He-Pin 2024-05-02
Sorry for now response, holidays here. I think we can including this in m1 and try with SWAR after that.